Moved save location to appdata, minor cleanup
Moved the code in Form1_Visible to Form1_Load and changed it to fire when
the form loads, instead of every time it was shown; the code in there is
only necessary once on load. Removed lines in SelectedHistoryItemChanged
setting defaults; data binding takes care of it. Fixed minor bug causing
multiple items in the history to be selected when one is removed.